+++ app/Models/ProductCategory.php
@@ -9,8 +9,6 @@ class ProductCategory extends Model
 {
     use HasFactory;
 
-    protected $guarded = [];
-
     public function products()
     {
         return $this->belongsToMany(Product::class, 'product_category_product', 'category_id', 'product_id', 'id')