+++ resources/js/Components/Notifications/NotificationCard.vue
@@ -1,0 +1,17 @@
+<template>
+            <div class="flex flex-row mb-6">
+                <div class="w-fit">
+                    IMAGE
+                </div>
+                <div class="ml-4 lg:ml-8">
+                    <h3 class="font-bold text-lg">Notification Title</h3>
+                    <p class="py-4">Notifications will go here.</p>
+                </div>
+            </div>
+</template>
+
+<script setup>
+
+
+
+</script>